Update Wheels (1).py
Haven't worked on this in a bit so added comments to clarify what I was attempting.  Current code is still extremely rough as my focus has been on resolving my inability to have one object modify another to result in a third object.  Once that is done, I will create modules and an effective method for users to exit.